refactor to get merge existing types with defaults
